+++ skillfarm/models/__init__.py
@@ -1,4 +1,5 @@
 from .characterskill import CharacterSkill
 from .general import General
-from .skillfarmaudit import SkillFarmAudit, SkillFarmSetup
+from .skillfarmaudit import SkillFarmAudit
+from .skillfarmsetup import SkillFarmSetup
 from .skillqueue import CharacterSkillqueueEntry
